Enhance data structures and utilities: Added Session and Trials classes, improved file handling in WidefieldData, and updated imports in __init__.pyi files. Refactored open/close methods for better state management in BaseData.
