+++ poulet_py/utils/__init__.pyi
@@ -1,5 +1,13 @@
 # ruff: noqa TID252
+from .analisys.common import Session, Trials
 from .experiment import ExperimentBlock, ExperimentRuntime, ExperimentTrial
 from .oscilloscope import Oscilloscope
 
-__all__ = ["ExperimentBlock", "ExperimentRuntime", "ExperimentTrial", "Oscilloscope"]
+__all__ = [
+    "ExperimentBlock",
+    "ExperimentRuntime",
+    "ExperimentTrial",
+    "Oscilloscope",
+    "Session",
+    "Trials",
+]
